Removes all dependencies to Handlebars and JQuery
The library din't depend to much on Handlebars. Components taking
a compiled Handlebars template are, in fact, able to work with any
function taking a model and returning HTML. So, why to say that
we need a 'handlebars template'? We are much more generic!

In addition, we removed a small dependency on JQuery, relying in
the standard querySelectorAll() function.