Fixes bug when adding event listeners to no-leaf nodes
When adding an event event listeners to components with a query
(e.g: #mylink) that match with html elements that are not leaf nodes
(e.g: the <a id="#mylink"> has an <img> inside), the component's root
event listener did not dispatch the event since the event.target
(i.e. the <img>) did not match with the query ("#mylink").

Now we match the event target (the <img>) to both "#mylink" and
"#mylink *" in order to dispatch clicks on the <img> as clicks on
its parent link.