Fixes missing exception when a renderer has more than one root
When a renderer generates html with more than one root, which is
illegal in fronty.js, an exception must be thrown. This commit
fixes this bug.

In addition a test is provided to test this behaviour. Moreover
a LogRecorded class is provided for tests interested in checking
log messages. This class is not used yet in any test, but it is
interesting for the future, because we are planning to show more
log messages to guide the user when programming with fronty.js.